+++ src/AlertsManagement/AlertsManagement/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 ## Upcoming Release
 
+* This version supports the new alert processing rule API (changed from action rule) 
+
 ## Version 0.3.0
 * Fixed bug for `Set-AzActionRule` when RecurrenceType is "Once" and no "ReccurentValue" provided [#14476]